Went through with Max and connected the refactor tests at current state. All customer tests pass
